act controller

+++ demondongo/src/main/java/com/mundongo/demondongo/controller/LugarTuristicoController.java
@@ -1,64 +1,69 @@
 package com.mundongo.demondongo.controller;
 
-import com.mundongo.demondongo.domain.LugarTuristico;
-import com.mundongo.demondongo.respository.LugarTuristicoRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import com.mundongo.demondongo.domain.LugarTuristico;
+import com.mundongo.demondongo.respository.LugarTuristicoRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/lugares-turisticos")
 public class LugarTuristicoController {
 
-    private final LugarTuristicoRepository lugarTuristicoRepository;
-
-    public LugarTuristicoController(LugarTuristicoRepository lugarTuristicoRepository) {
-        this.lugarTuristicoRepository = lugarTuristicoRepository;
-    }
+    @Autowired
+    private LugarTuristicoRepository lugarTuristicoRepository;
 
     @GetMapping
-    public List<LugarTuristico> getAllLugaresTuristicos() {
-        return lugarTuristicoRepository.findAll();
-    }
-
-    @PostMapping
-    public LugarTuristico createLugarTuristico(@RequestBody LugarTuristico lugarTuristico) {
-        return lugarTuristicoRepository.save(lugarTuristico);
+    public ResponseEntity<List<LugarTuristico>> read() {
+        List<LugarTuristico> query = lugarTuristicoRepository.findAll();
+        return new ResponseEntity<List<LugarTuristico>>(query, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<LugarTuristico> getLugarTuristicoById(@PathVariable Long id) {
-        return lugarTuristicoRepository.findById(id)
-               .map(lugar -> ResponseEntity.ok().body(lugar))
-               .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<LugarTuristico> readId(@PathVariable Long id) {
+        Optional<LugarTuristico> query = lugarTuristicoRepository.findById(id);
+        if (query.isPresent()) {
+            return new ResponseEntity<LugarTuristico>(query.get(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+    }
+
+    @PostMapping
+    public ResponseEntity<String> create(@RequestBody LugarTuristico lugarTuristico) {
+        lugarTuristicoRepository.save(lugarTuristico);
+        return new ResponseEntity<>("Lugar Turistico created :)))", HttpStatus.CREATED);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<LugarTuristico> updateLugarTuristico(@PathVariable Long id, @RequestBody LugarTuristico lugarTuristico) {
-        return lugarTuristicoRepository.findById(id)
-               .map(lugar -> {
-                   lugar.setNombre(lugarTuristico.getNombre());
-                   lugar.setDescripcion(lugarTuristico.getDescripcion());
-                   lugar.setDireccion(lugarTuristico.getDireccion());
-                   lugar.setLatitud(lugarTuristico.getLatitud());
-                   lugar.setLongitud(lugarTuristico.getLongitud());
-                   lugar.setUrlImagen(lugarTuristico.getUrlImagen());
-                   LugarTuristico updatedLugarTuristico = lugarTuristicoRepository.save(lugar);
-                   return ResponseEntity.ok().body(updatedLugarTuristico);
-               }).orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<String> update(@RequestBody LugarTuristico lugarTuristicoDetails, @PathVariable Long id) {
+        Optional<LugarTuristico> query = lugarTuristicoRepository.findById(id);
+        if (query.isPresent()) {
+            LugarTuristico lugar = query.get();
+            lugar.setNombre(lugarTuristicoDetails.getNombre());
+            lugar.setDescripcion(lugarTuristicoDetails.getDescripcion());
+            lugar.setDireccion(lugarTuristicoDetails.getDireccion());
+            lugar.setLatitud(lugarTuristicoDetails.getLatitud());
+            lugar.setLongitud(lugarTuristicoDetails.getLongitud());
+            lugar.setUrlImagen(lugarTuristicoDetails.getUrlImagen());
+            lugarTuristicoRepository.save(lugar);
+            return new ResponseEntity<>("Lugar Turistico updated :D", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Lugar Turistico not found :((", HttpStatus.NOT_FOUND);
+        }
     }
-
-    
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteLugarTuristico(@PathVariable Long id) {
-        return lugarTuristicoRepository.findById(id)
-               .map(lugar -> {
-                   lugarTuristicoRepository.deleteById(id);
-                   return ResponseEntity.ok().build();
-               }).orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        Optional<LugarTuristico> query = lugarTuristicoRepository.findById(id);
+        if (query.isPresent()) {
+            lugarTuristicoRepository.deleteById(id);
+            return new ResponseEntity<>("Lugar Turistico deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Lugar Turistico not found :((", HttpStatus.NOT_FOUND);
+        }
     }
-
-
 }
